started migrating functions from nfact_pp into nfact
