+++ NFACT/NFACT_PP/nfactpp_utils_functions.py
@@ -1,7 +1,7 @@
 import os
-import signal
 from datetime import datetime
 import glob
+from NFACT.utils.utils import error_and_exit
 
 
 def write_to_file(file_path: str, name: str, text: str) -> bool:
@@ -25,28 +25,6 @@ def write_to_file(file_path: str, name: str, text: str) -> bool:
         print(f"Unable to write to {file_path}/{name} due to :", e)
         return False
     return True
-
-
-def colours():
-    """
-    Function to print out text in colors
-
-    Parameters
-    ----------
-    None
-
-    Returns
-    -------
-    dict: dictionary object
-        dictionary of color strings
-    """
-    return {
-        "reset": "\033[0;0m",
-        "red": "\033[1;31m",
-        "pink": "\033[1;35m",
-        "purple": "\033[38;5;93m",
-        "darker_pink": "\033[38;5;129m",
-    }
 
 
 def make_directory(path: str) -> None:
@@ -90,70 +68,6 @@ def read_file_to_list(filename: str) -> list:
     with open(filename, "r") as file:
         lines = file.readlines()
     return [sub.rstrip() for sub in lines]
-
-
-def error_and_exit(bool_statement: bool, error_message=None):
-    """
-    Function to exit out of script
-    with error message if bool statement
-    is false
-    """
-    if not bool_statement:
-        if error_message:
-            col = colours()
-            print(col["red"] + error_message + col["reset"])
-        print("Exiting...\n")
-        exit(1)
-
-
-class Signit_handler:
-    """
-    A signit handler class. Will kill. Exits
-    programme safely.
-    """
-
-    def __init__(
-        self,
-    ) -> None:
-        self.register_handler()
-        self.suppress_messages = False
-
-    def register_handler(self) -> None:
-        """
-        Method to registers the SIGINT handler.
-        """
-        signal.signal(signal.SIGINT, self.handle_sigint)
-
-    def handle_sigint(self, sig, frame) -> None:
-        """
-        Method that handles the SIGINT signal (Ctrl+C)
-
-        Parameters
-        -----------
-        sig: The signal number
-        frame: The current stack frame
-        """
-        if not self.suppress_messages:
-            col = colours()
-            print(
-                f"\n{col['darker_pink']}Recieved kill signal (Ctrl+C). Terminating..."
-            )
-            print(f"Exiting...{col['reset']}\n")
-        exit(0)
-
-    @property
-    def get_suppress_messages(self):
-        """
-        Getter method for suppress_messages
-        """
-        return self.suppress_messages
-
-    @get_suppress_messages.setter
-    def set_suppress_messages(self, value: bool) -> None:
-        """
-        Setter for suppress_messages
-        """
-        self.suppress_messages = value
 
 
 def date_for_filename() -> str:
